+++ src/routes/+page.svelte
@@ -3,13 +3,14 @@
   import { listen } from "@tauri-apps/api/event";
   import * as remeda from "remeda";
   import { onMount } from "svelte";
-  import type { DeviceInfo } from "@vrlh/core";
+  import type { DeviceInfo, DeviceRemoteStatus } from "@vrlh/core";
   import { SvelteMap } from "svelte/reactivity";
-  import { slide } from "svelte/transition";
   import play from "$lib/icons/mingcute--play-fill.svg?raw";
   import pause from "$lib/icons/mingcute--pause-fill.svg?raw";
   import stop from "$lib/icons/mingcute--stop-fill.svg?raw";
   import { status } from "$lib/status.svelte";
+  import Device from "./device.svelte";
+  import Command from "./command.svelte";
 
   let pending = $state(true);
   const devices = new SvelteMap<string, DeviceInfo>();
@@ -39,11 +40,20 @@
     }
   }
 
-  function createOnclick(id: unknown, cmd: number): () => void {
+  const COMMAND_MAP = new Map<number, DeviceRemoteStatus>([
+    [0, "Stopped"],
+    [1, "Active"],
+    [2, "Standby"],
+  ]);
+  function createOnclick(cmd: number): () => void {
     return function onclick() {
-      invoke("power", { id, cmd }).catch((err: unknown) => {
-        status.push(JSON.stringify(err));
-      });
+      for (const device of devices.values()) {
+        const matcher = COMMAND_MAP.get(cmd);
+        if (matcher === device.remote) continue;
+        invoke("power", { id: device.id, cmd }).catch((err: unknown) => {
+          status.push(JSON.stringify(err));
+        });
+      }
     };
   }
 
@@ -57,7 +67,7 @@
 </script>
 
 <main class="p-2 space-y-2 min-h-0 flex-1 overflow-auto">
-  <div class="flex gap-4 items-center">
+  <div class="flex gap-2 items-center">
     <button
       class={[
         "px-4 py-2 bg-blue-900 b-(1 blue-950) rounded font-bold transition-colors",
@@ -69,66 +79,15 @@
     >
       Refresh
     </button>
+    <div class="flex b-(1 black) rounded divide-(x-1 black) overflow-hidden">
+      <Command onclick={createOnclick(0)} icon={stop} />
+      <Command onclick={createOnclick(2)} icon={pause} />
+      <Command onclick={createOnclick(1)} icon={play} />
+    </div>
   </div>
   <ul class="flex flex-col gap-2">
     {#each arr as device (device.addr)}
-      {@const { id, addr, name, local, remote } = device}
-      {@const disabled = local !== "Disconnected"}
-      <li
-        class="p-2 pl-3 flex justify-between b-(1 black) bg-neutral-800 rounded font-mono"
-        transition:slide
-      >
-        <div>
-          <div class="flex gap-2 items-center">
-            <h3 class="text-3xl font-bold">
-              {name}
-            </h3>
-            {#if disabled}
-              <div
-                class={[
-                  "h-2 w-2 rounded-full animate-pulse",
-                  local === "Initializing"
-                    ? "bg-yellow-800"
-                    : local === "Connected"
-                      ? "bg-blue-800"
-                      : "bg-red-800",
-                ]}
-                title={typeof device.local === "string"
-                  ? device.local
-                  : device.local && device.local.Error}
-              ></div>
-            {/if}
-          </div>
-          <div class="-mt-1 text-sm font-italic">
-            {addr}
-          </div>
-        </div>
-        <div
-          class="flex b-(1 black) rounded divide-(x-1 black) overflow-hidden"
-        >
-          {@render command({
-            device: id,
-            cmd: 0,
-            icon: stop,
-            disabled,
-            active: remote === "Stopped",
-          })}
-          {@render command({
-            device: id,
-            cmd: 2,
-            icon: pause,
-            disabled,
-            active: remote === "Standby",
-          })}
-          {@render command({
-            device: id,
-            cmd: 1,
-            icon: play,
-            disabled,
-            active: remote === "Active",
-          })}
-        </div>
-      </li>
+      <Device {device} />
     {/each}
     {#if !pending && devices.size === 0}
       <li class="p-4 b-(1 black) bg-neutral-800 rounded space-y-2">
@@ -140,25 +99,3 @@
 <footer class="px-4 py-1 bg-neutral-800 b-t-(1 black) font-italic">
   {status.current}
 </footer>
-
-{#snippet command(opts: {
-  device: unknown;
-  class?: string;
-  disabled?: boolean;
-  active?: boolean;
-  icon: string;
-  cmd: number;
-})}
-  <button
-    class={[
-      "h-full p-2 transition-colors bg-neutral-900 hover:bg-neutral-700",
-      "disabled:(bg-neutral-950 cursor-not-allowed data-[active]:bg-neutral-700)",
-      opts.class,
-    ]}
-    disabled={opts.disabled || opts.active || undefined}
-    data-active={opts.active || undefined}
-    onclick={createOnclick(opts.device, opts.cmd)}
-  >
-    {@html opts.icon}
-  </button>
-{/snippet}