+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add ability to set lighthouses to standby
+- Add ability to send command to all devices
 - Improve error handling and connection management
 - Fix window size issues happening on some systems